move old tests into new files
